backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
boto3
pdfminer.six
openai
python-multipart
python-dotenv
ratelimit
PyPDF2